# server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: NLM Service API
  description: API specification for nlm-service
  contact:
    email: support@nlmatics.com
  version: 2.2.13
servers:
  - url: http://localhost:5000/api
    description: NLMATICS Service API Endpoint
...
paths:
...
  /developerApiKey/accessToken:
    post:
      tags:
        - developerApiKey
      summary: access_token
      operationId: generate_developer_access_token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeveloperApiKey'
      responses:
        "200":
          description: Access token for the developer account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
      x-openapi-router-controller: server.controllers.developer_api_key_controller
      security: []
...
  /workspace/permission/{workspaceId}:
    post:
      tags:
        - workspace
      summary: share the workspace to an user with permission of "editor" or "viewer"
      operationId: add_user_permission
      parameters:
        - name: workspaceId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - name: email
          in: query
          required: true
          explode: false
          schema:
            type: string
        - name: permission
          in: query
          required: true
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: Returns the user id and permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdWithMessage'
      x-openapi-router-controller: server.controllers.workspace_controller
    get:
      tags:
        - workspace
      summary: return permission of an user
      operationId: get_user_permission
      parameters:
        - name: workspaceId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - name: email
          in: query
          required: true
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: Returns the user id and permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdWithMessage'
      x-openapi-router-controller: server.controllers.workspace_controller
...
  /promptLibrary:
    post:
      tags:
        - promptLibrary
      summary: Creates an entry in the prompt library.
      operationId: add_to_prompt_library
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Prompt"
      responses:
        "200":
          description: Returns the unique id assigned to the prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdWithMessage"
      x-openapi-router-controller: server.controllers.prompt_library_controller
    get:
      tags:
        - promptLibrary
      summary: Retrieves prompts matching the query parameters.
      operationId: get_prompts_for_workspace
      parameters:
        - name: workspaceId
          in: query
          required: true
          style: form
          explode: true
          schema:
            type: string
        - name: docId
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: string
        - name: promptType
          in: query
          required: false
          explode: true
          schema:
            type: string
        - name: queryScope
          in: query
          required: false
          explode: true
          schema:
            type: string
      responses:
        200:
          description: Returns list of prompts matching the query
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prompt'
                x-content-type: application/json
      x-openapi-router-controller: server.controllers.prompt_library_controller
...
security:
  - jwt: ['admin']
components:
  schemas:
...
    AccessToken:
      type: object
      properties:
        access_token:
          type: string
...
    DeveloperApiKey:
      type: object
      properties:
        api_key:
          type: string
        app_id:
          type: string
...
    IdWithMessage:
      type: object
      properties:
        id:
          type: string
        message:
          type: string
      example:
        id: id
        message: message
...
    Prompt:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        docId:
          type: string
        queryScope:
          type: string
        promptType:
          type: string
        title:
          type: string
        subtitle:
          type: string
        searchCriteria:
          type: object
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: server.auth.authenticate